add injection method
